requirements.txt
aiohttp
python-dotenv
openai
httpx
